+++ CHANGELOG.md
@@ -7,11 +7,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations
 on how to structure this file.
 This plugin uses [Semantic versioning](https://semver.org).
 
-## [2.9.0] - 2021-11-24
+## [2.9.0] - 2021-11-29
 
 ### Added
 
-- implemented Go to Definition
+- Implemented Go to Definition
+
+### Removed
+
+- OMT Declared import support
 
 ## [2.8.8] - 2021-11-23